[bom-bom-study#7][ADD] Family 테이블 추가 및 다대다 관계 CRUD service 작성

+++ src/main/java/me/jae57/woodywoody/model/Scent.java
@@ -5,7 +5,7 @@ import lombok.*;
 @Builder
 @Getter
 public class Scent {
-    private int scentId;
+    private Long scentId;
     private String scentName;
     private String scentKorName;
     private String brand;
